Functions

IDFDataCanada.get_idf — Function

get_idf(fileName::String)

This function reads ECCC IDF text files and returns station infos (ID, latitude, longitude, altitude, and station name) and a DataFrame containing observed annual maximum in mm (Table 1) for different durations.

IDFDataCanada.txt2csv — Function

txt2csv(input_dir::String, output_dir::String)

This function returns CSV files of observed annual maximum for each station and one CSV file containing all station info (name, province, ID, lat, lon, elevation, number of years, data CSV filenames, original filenames) for a province.

IDFDataCanada.txt2netcdf — Function

txt2netcdf(input_dir::String, output_dir::String)

This function returns netCDF files containing observed annual maximum data and station info for each station of a province.

IDFDataCanada.data_download — Function

data_download(province::String, output_dir::String, format::String; split::Bool, rm_temp::Bool)

This function downloads IDF data from ECCC Google Drive directory for a province and generates CSV or netCDF files. CSV format is selected by default.

data_download(province::Array{String}, output_dir::String, url::String, file_basename::String)

This function downloads IDF data from ECCC Google Drive directory for multiple provinces and generates CSV or netCDF files. CSV format is selected by default.

IDFDataCanada.netcdf_generator — Function

netcdf_generator(fileName::String)

This functions generates empty netCDF files (used by txt2netcdf).